Move cached_state.rs into state directory.
