+++ blockifier/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod cached_state;
 pub mod execution;
 pub mod state;
 pub mod test_utils;
