+++ blockifier/src/state.rs
@@ -1,2 +1,3 @@
+pub mod cached_state;
 pub mod errors;
 pub mod state_reader;
